+++ packages/tailwindcss/src/utilities.ts
@@ -254,7 +254,7 @@ function resolveThemeColor<T extends ThemeKey>(
       break
     }
     case 'current': {
-      value = 'currentColor'
+      value = 'currentcolor'
       break
     }
     default: {
@@ -5439,7 +5439,7 @@ export function createUtilities(theme: Theme) {
 
     staticUtility('ring-inset', [boxShadowProperties, ['--tw-ring-inset', 'inset']])
 
-    let defaultRingColor = theme.get(['--default-ring-color']) ?? 'currentColor'
+    let defaultRingColor = theme.get(['--default-ring-color']) ?? 'currentcolor'
     function ringShadowValue(value: string) {
       return `var(--tw-ring-inset,) 0 0 0 calc(${value} + var(--tw-ring-offset-width)) var(--tw-ring-color, ${defaultRingColor})`
     }
@@ -5515,7 +5515,7 @@ export function createUtilities(theme: Theme) {
     ])
 
     function insetRingShadowValue(value: string) {
-      return `inset 0 0 0 ${value} var(--tw-inset-ring-color, currentColor)`
+      return `inset 0 0 0 ${value} var(--tw-inset-ring-color, currentcolor)`
     }
     utilities.functional('inset-ring', (candidate) => {
       if (!candidate.value) {